Use flyway to manage the schema
Hibernate is creating the schema for us because we're using an embedded
database. As soon as we switch to a production database, nothing happens
as you probably want to manage the schema yourself.

We can demonstrate that by adding the MySQL driver, a `local` profile
pointing to a local database and start our app with that profile on.

Now, let's ask Hibernate to validate the schema and use flyway to
initialize the schema. Again, we're using a version to initialize our
speaker table for demo purposes. You can use a separate profile in dev
to do that only. Note also that because we have devtools, the h2 console
is available automatically locally.

It's time to put our app in the cloud. It's using H2 by default as we're
not using any service yet. We'll fix that in the next demo