Env customizations demo
For demonstrating purposes we create a simple key for our app to change
the hello message of the home controller. By adding the annotation
processor, we also easily integrate in our IDE and have now many places
where we can tune that property:

* In any configuration file (including profiles)
* In the run config
* On the command line
* Using system properties or environment variables

We can now start our app on the command-line and tune it. Then we can
push it to PWS to showcase how the `cf set-env` command can tune our
app.