Enable actuator endpoints
By default, only info and status are exposed. This is largely due to the
fact that the default security gives you access to everything once you
are authenticated.

We can see our `HERO` role is indeed ignored so we'll have to fix that
and replace the property we've removed in the previous step.