Custom security configuration
If a custom security configuration is provided, it completely replaces
what Spring Boot does by default. In Spring Boot 2, several new helper
classes were introduced to quickly allow you to replicate part of what
the auto-configuration does.

Note that even if we changed the path of the health endpoint, we simply
refer to its.